Move host floating point environment handling in host.h/host.cc
