+++ lib/evaluate/intrinsics-library.cc
@@ -12,20 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// This file defines host runtimes functions that can be used for folding
+// This file defines host runtime functions that can be used for folding
 // intrinsic functions.
 // The default HostIntrinsicProceduresLibrary is built with <cmath> and
 // <complex> functions that are guaranteed to exist from the C++ standard.
 
 #include "intrinsics-library-templates.h"
-#include "../common/idioms.h"
-#include <cerrno>
-#include <cfenv>
-#include <sstream>
+#include <cmath>
+#include <complex>
 
 namespace Fortran::evaluate {
 
-using namespace Fortran::parser::literals;
 // Note: argument passing is ignored in equivalence
 bool HostIntrinsicProceduresLibrary::HasEquivalentProcedure(
     const IntrinsicProcedureRuntimeDescription &sym) const {
@@ -53,8 +50,8 @@ bool HostIntrinsicProceduresLibrary::HasEquivalentProcedure(
 
 // Map numerical intrinsic to  <cmath>/<complex> functions
 
-// TODO mapping to <cmath> function to be tested.<cmath> func takes
-// real arg for n
+// C++ Bessel functions take a floating point as first argument.
+// Fortran Bessel functions take an integer.
 template<typename HostT> static HostT Bessel_jn(std::int64_t n, HostT x) {
   return std::cyl_bessel_j(static_cast<HostT>(n), x);
 }
@@ -103,7 +100,7 @@ void AddLibmComplexHostProcedure(
   }
 }
 
-// Defines which host runtime functions will be used for folding
+// Define which host runtime functions will be used for folding
 
 void HostIntrinsicProceduresLibrary::DefaultInit() {
 
@@ -114,94 +111,5 @@ void HostIntrinsicProceduresLibrary::DefaultInit() {
   AddLibmComplexHostProcedure<float>(*this);
   AddLibmComplexHostProcedure<double>(*this);
   AddLibmComplexHostProcedure<long double>(*this);
-}
-
-void HostFloatingPointEnvironment::SetUpHostFloatingPointEnvironment(
-    FoldingContext &context) {
-  errno = 0;
-  if (feholdexcept(&originalFenv_) != 0) {
-    common::die("Folding with host runtime: feholdexcept() failed: %s",
-        std::strerror(errno));
-    return;
-  }
-  if (fegetenv(&currentFenv_) != 0) {
-    common::die("Folding with host runtime: fegetenv() failed: %s",
-        std::strerror(errno));
-    return;
-  }
-#if __x86_64__
-  if (context.flushSubnormalsToZero()) {
-    currentFenv_.__mxcsr |= 0x8000;  // result
-    currentFenv_.__mxcsr |= 0x0040;  // operands
-  } else {
-    currentFenv_.__mxcsr &= ~0x8000;  // result
-    currentFenv_.__mxcsr &= ~0x0040;  // operands
-  }
-#else
-  // TODO other architectures
-  context.messages().Say(
-      "TODO: flushing mode for subnormals is not set for this host architecture when folding with host runtime functions"_en_US);
-#endif
-  errno = 0;
-  if (fesetenv(&currentFenv_) != 0) {
-    common::die("Folding with host runtime: fesetenv() failed: %s",
-        std::strerror(errno));
-    return;
-  }
-  switch (context.rounding().mode) {
-  case RoundingMode::TiesToEven: fesetround(FE_TONEAREST); break;
-  case RoundingMode::ToZero: fesetround(FE_TOWARDZERO); break;
-  case RoundingMode::Up: fesetround(FE_UPWARD); break;
-  case RoundingMode::Down: fesetround(FE_DOWNWARD); break;
-  case RoundingMode::TiesAwayFromZero:
-    fesetround(FE_TONEAREST);
-    context.messages().Say(
-        "TiesAwayFromZero rounding mode is not available not available when folding constants with host runtime. Using TiesToEven instead."_en_US);
-    break;
-  }
-  errno = 0;
-}
-void HostFloatingPointEnvironment::CheckAndRestoreFloatingPointEnvironment(
-    FoldingContext &context) {
-  int errnoCapture{errno};
-  int exceptions{fetestexcept(FE_ALL_EXCEPT)};
-  RealFlags flags;
-  if (exceptions & FE_INVALID) {
-    flags.set(RealFlag::InvalidArgument);
-  }
-  if (exceptions & FE_DIVBYZERO) {
-    flags.set(RealFlag::DivideByZero);
-  }
-  if (exceptions & FE_OVERFLOW) {
-    flags.set(RealFlag::Overflow);
-  }
-  if (exceptions & FE_UNDERFLOW) {
-    flags.set(RealFlag::Underflow);
-  }
-  if (exceptions & FE_INEXACT) {
-    flags.set(RealFlag::Inexact);
-  }
-
-  if (flags.empty()) {
-    if (errnoCapture == EDOM) {
-      flags.set(RealFlag::InvalidArgument);
-    }
-    if (errnoCapture == ERANGE) {
-      // can't distinguish over/underflow from errno
-      flags.set(RealFlag::Overflow);
-    }
-  }
-
-  if (!flags.empty()) {
-    RealFlagWarnings(context, flags, "folding function with host runtime");
-  }
-  errno = 0;
-  if (fesetenv(&originalFenv_) != 0) {
-    std::fprintf(stderr, "fesetenv() failed: %s\n", std::strerror(errno));
-    common::die(
-        "Folding with host runtime: fesetenv() failed while restoring fenv: %s",
-        std::strerror(errno));
-  }
-  errno = 0;
 }
 }
